requirements.txt
torch
torchvision
pillow
kornia
transformers